+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Using another Bias detection model which works better on different devices [valurank/distilroberta-bias](https://huggingface.co/valurank/distilroberta-bias)
 - Updated the roadmap
+- `BanSubstrings` can redact found substrings
 
 ### Removed
 - Usage of `accelerate` library for inference. Instead, it will detect device using `torch`